backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ryan Loh <ryanloh29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mem0ai = "^0.1.21"
uvicorn = "^0.32.0"
fastapi = "^0.115.2"
python-dotenv = "^1.0.1"
supabase = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
